% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRB_long}
\alias{CRB_long}
\title{Long Format Fish Returns at Bonneville Dam}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 783 rows and 3 columns.
}
\usage{
CRB_long
}
\description{
A long dataset containing annual counts of fish returns at Bonneville Dam in
the Columbia River Basin.
}
\keyword{datasets}
