% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hkitty_palettes}
\alias{hkitty_palettes}
\title{Complete list of palettes}
\usage{
hkitty_palettes
}
\value{
No return value, called for side effects (list of color palettes).
}
\description{
Complete list of palettes
}
