% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hkitty_palette}
\alias{hkitty_palette}
\title{A Hello kitty palette generator}
\usage{
hkitty_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{hellokitty1}, \code{hellokitty2}}

\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from Wes Anderson movies.
}
\examples{
hkitty_palette("hellokitty1")
hkitty_palette("hellokitty2")
hkitty_palette("hellokitty1", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- hkitty_palette(21, name = "hellokitty2", type = "continuous")
}
\keyword{colors}
